requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cmake==0.9.0
cryptography==2.2.2
cycler==0.10.0
dlib==19.9.0
face-recognition==1.2.1
face-recognition-models==0.3.0
Flask==0.12.2
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
matplotlib==2.0.2
numpy==1.14.0
opencv-python==3.3.0.10
Pillow==5.0.0
ptyprocess==0.5.2
pybase64==0.2.1
pycparser==2.18
pycryptodome==3.6.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
scikit-learn==0.19.0
scipy==1.0.0
urllib3==1.22

